fetch proj info
