+++ dds_cli/dds_gui/models/project.py
@@ -23,7 +23,6 @@ class Project:
             project_content=project_content,
         )
 
-
 @dataclass
 class ProjectContentData:
     """A project content."""
